pyproject.toml
[tool.poetry]
name = "githublfs"
version = "0.1.3"
description = "Simple library to commit lfs files to Github repos"
authors = ["Anès Foufa <anes.foufa@upply.com>"]
readme = "readme.rst"
classifiers = [
    "Topic :: Software Development :: Version Control",
    "Topic :: Software Development :: Version Control :: Git",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
keywords = [
    "Github",
    "LFS",
]
homepage = "https://github.com/AnesFoufa/githublfs"

[tool.poetry.dependencies]
python = "^3.7"
PyGithub = "^1.55"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
mypy = "^0.971"
black = "^22.6.0"
twine = "^4.0.1"
requests-mock = "^1.9.3"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
